+++ src/collective/themesitesetup/interfaces.py
@@ -7,6 +7,7 @@ DEFAULT_ENABLED_PROFILE_NAME = 'install'
 DEFAULT_DISABLED_PROFILE_NAME = 'uninstall'
 DEFAULT_ENABLED_LOCALES_NAME = 'locales'
 DEFAULT_ENABLED_MODELS_NAME = 'models'
+DEFAULT_ENABLED_RESOURCES_NAME = 'resources'
 
 NO = ['off', 'false', 'no', '0']
 YES = ['on', 'true', 'yes', '1']