+++ CHANGES.rst
@@ -1,13 +1,7 @@
 Changelog
 =========
 
-1.2.2 (unreleased)
-------------------
-
-- Nothing changed yet.
-
-
-1.2.1 (2016-09-01)
+1.3.2 (unreleased)
 ------------------
 
 - Fix issue where message catalog support allowed (mostly accidentally)
@@ -15,11 +9,31 @@ Changelog
   now ignored.
   [datakurre]
 
+
+1.3.1 (2016-12-14)
+------------------
+
+- Add to purge plone.app.blocks' site layout cache after resource directory
+  copy
+  [datakurre]
+
+
+1.3.0 (2016-11-22)
+------------------
+
+- Add support for populating persistent (plone.resource) resource directories
+  [datakurre]
+
+- Refactor permission support to use zope.app.localpermission
+  [datakurre]
+
+
 1.2.0 (2016-08-17)
 ------------------
 
 - Add support for TTW custom permissions
   [datakurre]
+
 
 1.1.0 (2016-08-12)
 ------------------
@@ -28,6 +42,7 @@ Changelog
   from ``./models/Xxxxxx.xml``
   [datakurre]
 
+
 1.0.1 (2016-08-11)
 ------------------
 
@@ -35,12 +50,14 @@ Changelog
   catalog
   [datakurre]
 
+
 1.0.0 (2016-08-11)
 ------------------
 
 - Add support for registering i18n message catalogs directly from theme
   from ``./locales/xx/LC_MESSAGES/yyyyy.po``
   [datakurre]
+
 
 0.13.0 (2015-04-23)
 -------------------
@@ -54,8 +71,10 @@ Changelog
 
 - Move custom GS import export adapters to external configuration
   [datakurre]
+
 - Fix to register setup forms for p.a.theming layer
   [datakurre]
+
 
 0.11.1 (2015-04-04)
 -------------------
@@ -63,15 +82,18 @@ Changelog
 - Update README
   [datakurre]
 
+
 0.11.0 (2015-04-04)
 -------------------
 
 - Add site setup import view to allow testing or manual upgrading of site
   setups
   [datakurre]
+
 - Add option to disable setup steps import via plugin configuration variable in
   theme manifest (either with ``enabled = false`` or ``disabled = true``)
   [datakurre]
+
 
 0.10.0 (2015-04-03)
 -------------------
@@ -84,6 +106,7 @@ Changelog
 - Add GS content export/import to support non-CMF-containers, PythonScripts
   and PageTemplates.
   [datakurre]
+
 
 0.9.0 (2015-04-01)
 ------------------